+++ react-app/src/components/DashboardPage/index.js
@@ -1,9 +1,10 @@
 import LeftNavigationBar from "../LeftNavigationBar";
-
+import TopNavigation from "../TopNavigationBar";
 function DashboardPage() {
     return (
         <>
             <h1>Dashboard Page</h1>
+            <TopNavigation/>
             <LeftNavigationBar />
         </>
     );
