+++ react-app/src/components/Navigation/index.js
@@ -1,70 +1,24 @@
-// import React from 'react';
-// import { NavLink } from 'react-router-dom';
-// import { useSelector } from 'react-redux';
-// import ProfileButton from './ProfileButton';
-// import './Navigation.css';
+import React from 'react';
+import { NavLink } from 'react-router-dom';
+import { useSelector } from 'react-redux';
+import ProfileButton from './ProfileButton';
+import './Navigation.css';
 
-// function Navigation({ isLoaded }){
-// 	const sessionUser = useSelector(state => state.session.user);
+function Navigation({ isLoaded }){
+	const sessionUser = useSelector(state => state.session.user);
 
-// 	return (
-// 		<ul>
-// 			<li>
-// 				<NavLink exact to="/">Home</NavLink>
-// 			</li>
-// 			{isLoaded && (
-// 				<li>
-// 					<ProfileButton user={sessionUser} />
-// 				</li>
-// 			)}
-// 		</ul>
-// 	);
-// }
-
-// export default Navigation;
-
-
-
-
-
-
-import React from "react";
-import { NavLink } from "react-router-dom";
-import { useSelector } from "react-redux";
-import ProfileButton from "./ProfileButton";
-import "./Navigation.css";
-
-
-function Navigation({ isLoaded }) {
-  const sessionUser = useSelector((state) => state.session.user);
-  const headerLogin = sessionUser?"header-login":""
-  const titleLogin = sessionUser?"title-login":""
-  return (
-
-    <header className={`header ${headerLogin}`}>
-      <div className="container">
-		<nav>
-		<ul className="navigation">
-          <li>
-            <NavLink className="logo" exact to="/">
-              <img
-                id="logo" alt="split-smart icon"
-                src="https://res.cloudinary.com/dr1ekjmf4/image/upload/v1687712495/images_lhirzh.png"
-              />
-              <h1 className={`title ${titleLogin}`}>SplitSmart</h1>
-            </NavLink>
-          </li>
-          {isLoaded && (
-            <li>
-              <ProfileButton user={sessionUser} />
-            </li>
-          )}
-        </ul>
-		</nav>
-       
-      </div>
-    </header>
-  );
+	return (
+		<ul>
+			<li>
+				<NavLink exact to="/">Home</NavLink>
+			</li>
+			{isLoaded && (
+				<li>
+					<ProfileButton user={sessionUser} />
+				</li>
+			)}
+		</ul>
+	);
 }
 
 export default Navigation;
