+++ react-app/src/components/TopNavigationBar/Navigation.css
@@ -1,0 +1,132 @@
+.profile-dropdown {
+    position: absolute;
+   
+    background-color: #fff;
+    top: 100px;
+    right: 50px;
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    border: 1px solid #000;
+    box-shadow: 5px 5px 5px #000;
+    padding: 15px;
+    align-items: flex-start;
+    justify-content: flex-start;
+  }
+  
+  .hidden {
+    display: none;
+  }
+  
+  
+  .header-login{
+    background-color: #1CC29F;
+  }
+  
+  .header{
+   
+    height: 100px;
+    width: 100%;
+    position: fixed;
+  }
+  .header img{
+    height: 70px;
+    width: 70px;
+  }
+  .nav-btn img{
+    height: 72px;
+    width: 72px;
+    border-radius: 50%;
+  
+  }
+  .logo{
+    display: flex;
+    align-items: center;
+  }
+  
+  .title{
+    font-size: 20px;
+    text-transform: uppercase;
+    font-weight: 700;
+    margin-left: 15px;
+    color: #000;
+    
+  }
+  
+  
+  .title-login{
+    color: #fff;
+  }
+  
+  .navigation{
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin: 0;
+    height: 100px;
+  }
+  
+  
+  .navigation-menu{
+    display: flex;
+    gap:10px
+  }
+  
+  
+  
+  .nav-login{
+    color: #1CC29F;
+  }
+  
+  
+  .nav-signup{
+    color: #fff;
+    background-color: #1CC29F;
+    border-radius: 4px;
+  }
+  
+  
+  .navigation-link{
+    font-size: 18px;
+    padding: 15px;
+  }
+  
+  
+  
+  .nav-btn{
+    border-radius: 50%;
+    width: 70px;
+    height: 70px;
+    border: none;
+    padding: 0;
+    margin: 0;
+  }
+  
+  
+  ul{
+    list-style: none;
+    
+  }
+  .container{
+    max-width: 1200px;
+    margin: 0 auto;
+  }
+  
+  a{
+    text-decoration: none;
+  }
+  
+  
+  
+  .log-out{
+    padding: 10px;
+    border: none;
+    border-radius: 3px;
+    
+  }
+  
+  .log-out:hover{
+    background-color: #1CC29F;
+    color: #fff;
+    
+  }
