+++ react-app/src/components/TopNavigationBar/ProfileButton.js
@@ -1,9 +1,7 @@
 import React, { useState, useEffect, useRef } from "react";
 import { useDispatch } from "react-redux";
 import { logout } from "../../store/session";
-import OpenModalButton from "../OpenModalButton";
-import LoginFormModal from "../LoginFormModal";
-import SignupFormModal from "../SignupFormModal";
+import { NavLink } from "react-router-dom";
 
 function ProfileButton({ user }) {
   const dispatch = useDispatch();
@@ -17,15 +15,12 @@ function ProfileButton({ user }) {
 
   useEffect(() => {
     if (!showMenu) return;
-
     const closeMenu = (e) => {
       if (!ulRef.current.contains(e.target)) {
         setShowMenu(false);
       }
     };
-
     document.addEventListener("click", closeMenu);
-
     return () => document.removeEventListener("click", closeMenu);
   }, [showMenu]);
 
@@ -33,40 +28,40 @@ function ProfileButton({ user }) {
     e.preventDefault();
     dispatch(logout());
   };
-
   const ulClassName = "profile-dropdown" + (showMenu ? "" : " hidden");
-  const closeMenu = () => setShowMenu(false);
-
+ 
   return (
     <>
-      <button onClick={openMenu}>
-        <i className="fas fa-user-circle" />
-      </button>
-      <ul className={ulClassName} ref={ulRef}>
+      <ul className="navigation-menu" ref={ulRef}>
         {user ? (
           <>
-            <li>{user.username}</li>
-            <li>{user.email}</li>
-            <li>
-              <button onClick={handleLogout}>Log Out</button>
-            </li>
+            <button className="nav-btn" onClick={openMenu}>
+              <img src={user.image_url} alt={user.name} />
+            </button>
+            <ul className={ulClassName} >
+              <li> Hello, {user.name}!</li>
+             
+              <li>
+                <button className="log-out" onClick={handleLogout}>Log Out</button>
+              </li>
+            </ul>
           </>
         ) : (
           <>
-            <OpenModalButton
-              buttonText="Log In"
-              onItemClick={closeMenu}
-              modalComponent={<LoginFormModal />}
-            />
-
-            <OpenModalButton
-              buttonText="Sign Up"
-              onItemClick={closeMenu}
-              modalComponent={<SignupFormModal />}
-            />
+            <li>
+              <NavLink to="/login" className="navigation-link nav-login">
+                Log in
+              </NavLink>
+            </li>
+            <li>
+              <NavLink to="/signup" className="navigation-link nav-signup">
+                Sign up
+              </NavLink>
+            </li>
           </>
         )}
       </ul>
+     
     </>
   );
 }
